% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPreferenceOrder}
\alias{checkPreferenceOrder}
\title{Check if preference order is complete}
\usage{
checkPreferenceOrder(pref)
}
\arguments{
\item{pref}{is a matrix with a preference ordering}
}
\value{
a matrix with preference orderings with proper C++ indices or NULL
if the preference order is not complete.
}
\description{
This function checks if a given preference ordering is complete. If needed
it transforms the indices from R indices (starting at 1) to C++ indices
(starting at zero).
}

