% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkStability}
\alias{checkStability}
\title{Check if a matching is stable}
\usage{
checkStability(uWorkers, uFirms, proposals, engagements)
}
\arguments{
\item{uWorkers}{is a matrix with cardinal utilities of the proposing side of the
market}

\item{uFirms}{is a matrix with cardinal utilities of the courted side of the
market}

\item{proposals}{is a matrix that contains the id of the firm that a given
worker is matched to: the first row contains the id of the firm that is
matched with the first worker, the second row contains the id of the firm
that is matched to the second worker, etc. The column dimension accommodates
the scenario when there are multiple workers of the same type. These must be
given in R style indexing.}

\item{engagements}{is a matrix that contains the id of the worker that a given
firm is matched to. The column dimension accommodates multi-worker firms. These
must be given in R style indexing.}
}
\value{
true if the matching is stable, false otherwise
}
\description{
This function checks if a given matching is stable for a particular set of
preferences. This function can check if a given check one-to-one,
one-to-many, or many-to-one matching is stable.
}

