% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkStabilityRoommate}
\alias{checkStabilityRoommate}
\title{Check if a matching solves the stable roommate problem}
\usage{
checkStabilityRoommate(pref, matchings)
}
\arguments{
\item{pref}{is a matrix with ordinal rankings of the participants}

\item{matchings}{is an nx1 matrix encoding who is matched to whom using
R style indexing}
}
\value{
true if the matching is stable, false otherwise
}
\description{
This function checks if a given matching is stable for a particular set of
preferences. This function checks if there's an unmatched pair that would
rather be matched with each other than with their assigned partners.
}

