% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/galeshapley.R
\name{galeShapley.marriageMarket}
\alias{galeShapley.marriageMarket}
\title{Computes Gale-Shapley algorithm to find solution to the stable marriage
problem}
\usage{
galeShapley.marriageMarket(proposerUtils = NULL, reviewerUtils = NULL,
  proposerPref = NULL, reviewerPref = NULL)
}
\arguments{
\item{proposerUtils}{is a matrix with cardinal utilities of the proposing
side of the market. If there are \code{n} proposers and \code{m} reviewers
in the market, then this matrix will be of dimension \code{m} by \code{n}.
The \code{i,j}th element refers to the payoff that individual \code{j}
receives from being matched to individual \code{i}.}

\item{reviewerUtils}{is a matrix with cardinal utilities of the courted side
of the market. If there are \code{n} proposers and \code{m} reviewers
in the market, then this matrix will be of dimension \code{n} by \code{m}.
The \code{i,j}th element refers to the payoff that individual \code{j}
receives from being matched to individual \code{i}.}

\item{proposerPref}{is a matrix with the preference order of the proposing
side of the market (only required when \code{proposerUtils} is not
provided). If there are \code{n} proposers and \code{m} reviewers
in the market, then this matrix will be of dimension \code{m} by \code{n}.
The \code{i,j}th element refers to the ID of individual \code{j}'s
\code{i}th most favorite partner.}

\item{reviewerPref}{is a matrix with the preference order of the courted side
of the market (only required when \code{reviewerUtils} is not provided). If
there are \code{n} proposers and \code{m} reviewers in the market, then
this matrix will be of dimension \code{n} by \code{m}. The \code{i,j}th
element refers to the ID of individual \code{j}'s \code{i}th most favorite
partner.}
}
\value{
A list with the successful proposals and engagements:
  \code{proposals} is a vector whose nth element contains the id of the
  reviewer that proposer n is matched to. \code{engagements} is a vector
  whose nth element contains the id of the proposer that reviewer n is
  matched to. \code{single.proposers} is a vector that lists the ids of
  remaining single proposers. \code{single.reviewers} is a vector that lists
  the ids of remaining single reviewers.
}
\description{
This function returns the one-to-one matching. The function needs some
description of individuals preferences as inputs. That can be in the form of
cardinal utilities or preference orders (or both).
}
\examples{
nmen = 5
nwomen = 4
uM = matrix(runif(nmen*nwomen), nrow = nwomen, ncol = nmen)
uW = matrix(runif(nwomen*nmen), nrow = nmen, ncol = nwomen)
results = galeShapley.marriageMarket(uM, uW)

prefM = sortIndex(uM)
prefW = sortIndex(uW)
results = galeShapley.marriageMarket(proposerPref = prefM, reviewerPref = prefW)
}

