% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\docType{package}
\name{matchingR-package}
\alias{matchingR-package}
\title{matchingR: Efficient Computation of Matching Algorithms in R
  and C++}
\description{
matchingR is an R Package that efficiently computes matching
  algorithms for large scale matching markets. It currently implements the
  Gale-Shapley Algorithm for two-sided matching markets and Irving's
  Algorithm for one-sided matching markets. This package can be useful when
  the number of market participants is large or when very many matchings need
  to be computed (e.g. for extensive simulations or for estimation purposes).
  The package has successfully been used to simulate preferences and compute
  the matching with 30,000 participants on each side of the market. The
  package provides functions to compute the solution to the
  \href{http://en.wikipedia.org/wiki/Stable_matching}{stable marriage
  problem}, to the
  \href{http://en.wikipedia.org/wiki/Hospital_resident}{college admission
  problem}, and to the
  \href{https://en.wikipedia.org/wiki/Stable_roommates_problem}{stable
  roommates problem}
}
\examples{
# stable marriage problem
nmen = 25
nwomen = 20
uM = matrix(runif(nmen*nwomen), nrow=nwomen, ncol=nmen)
uW = matrix(runif(nwomen*nmen), nrow=nmen, ncol=nwomen)
results = one2one(uM, uW)
checkStability(uM, uW, results$proposals, results$engagements)

# college admissions problem
nstudents = 25
ncolleges = 5
uStudents = matrix(runif(nstudents*ncolleges), nrow=ncolleges, ncol=nstudents)
uColleges = matrix(runif(nstudents*ncolleges), nrow=nstudents, ncol=ncolleges)
results = one2many(uStudents, uColleges, slots=4)
checkStability(uStudents, uColleges, results$proposals, results$engagements)

# stable roommate problem
N = 10
u = matrix(runif(N^2),  nrow = N, ncol = N)
results = onesided(utils = u)
}
\author{
Jan Tilly, Nick Janetos
}
\references{
Gale, D. and Shapley, L.S. (1962). College admissions and the
  stability of marriage. \emph{The American Mathematical Monthly}, 69(1):
  9--15.

Irving, R. W. (1985). An efficient algorithm for the "stable
  roommates" problem. \emph{Journal of Algorithms}, 6(4): 577--595
}

