% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{one2many}
\alias{one2many}
\title{Compute the one-to-many matching}
\usage{
one2many(proposerUtils = NULL, reviewerUtils = NULL, proposerPref = NULL,
  reviewerPref = NULL, slots = 1)
}
\arguments{
\item{proposerUtils}{is a matrix with cardinal utilities of the proposing
side of the market}

\item{reviewerUtils}{is a matrix with cardinal utilities of the courted side
of the market}

\item{proposerPref}{is a matrix with the preference order of the proposing
side of the market (only required when \code{uM} is not provided)}

\item{reviewerPref}{is a matrix with the preference order of the courted
side of the market (only required when \code{prefW} is not provided)}

\item{slots}{is an integer with the number slots per firm}
}
\value{
A list with the successful proposals and engagements:
\code{proposals} is a vector whose nth element contains the id of the firm
that worker n is matched to.
\code{engagements} is a matrix whose nth row contains the ids of the workers
that firm n is matched to.
\code{single.proposers} is a vector that lists the ids of remaining unmatched
workers
\code{single.reviewers} is a vector that lists the ids of remaining vacant
firms (if a firm is two vacancies it will be listed twice)
}
\description{
This function returns the one-to-many matching. Workers are making proposals.
The function needs some description of individuals preferences as inputs.
That can be in the form of cardinal utilities or preference orders (or both).
}

