% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{onesided}
\alias{onesided}
\title{Compute matching for one-sided markets}
\usage{
onesided(pref = NULL)
}
\arguments{
\item{pref}{An n-1xn matrix, with each column representing the cardinal
utilities of each agent over matches with the other agents, so that, e.g.,
if element (4, 6) of this matrix is 2, then agent 6 ranks agent 2 4th.}
}
\value{
A list of length n corresponding to the matchings being made, so that
e.g. if the 4th element is 6 then agent 4 was matched with agent 6.
}
\description{
This function returns a stable roommate matching for a one-sided market
using Irving (1985)'s algorithm. Stable matchings are neither guaranteed
to exist, nor to be unique.
}
\examples{
test = matrix(c(3, 4, 2, 6, 5,
                6, 5, 4, 1, 3,
                2, 4, 5, 1, 6,
                5, 2, 3, 6, 1,
                3, 1, 2, 4, 6,
                5, 1, 3, 4, 2), nrow=5, ncol=6);
results = onesided(test)
}

