% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Make a package environmental variable with the storage order}
\format{\preformatted{<environment: 0x483dde0> 
}}
\usage{
pkg.env
}
\description{
\code{pkg.env} is a package environment that contains the variable
\code{column.major} that indicates if preferences are stored in column
major order (default) or row major order.
}
\keyword{datasets}

