% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.row}
\alias{rep.row}
\title{Repeat each row of a matrix n times}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{is the input matrix}

\item{n}{is the number of repetitions}
}
\value{
matrix with repeated rows
}
\description{
This function repeats each row of a matrix n times
}

