% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{repcol}
\alias{repcol}
\title{Repeat each column of a matrix n times}
\usage{
repcol(x, n)
}
\arguments{
\item{x}{is the input matrix}

\item{n}{is the number of repetitions}
}
\value{
matrix with repeated columns
}
\description{
This function repeats each column of a matrix n times
}

