% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{reprow}
\alias{reprow}
\title{Repeat each row of a matrix n times}
\usage{
reprow(x, n)
}
\arguments{
\item{x}{is the input matrix}

\item{n}{is the number of repetitions}
}
\value{
matrix with repeated rows
}
\description{
This function repeats each row of a matrix n times
}

