% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roommate.R
\name{roommate.checkStability}
\alias{roommate.checkStability}
\title{Check if a roommate matching is stable}
\usage{
roommate.checkStability(pref, matchings)
}
\arguments{
\item{pref}{An n-1xn matrix, with each column representing the ordinal
utilities of each agent over matches with the other agents, so that, e.g.,
if element (4, 6) of this matrix is 2, then agent 4 ranks agent 2 6th. The
matrix accepts either 0-based indexing (C++ style) or 1-based indexing (R
style).}

\item{matchings}{is a matrix with matchings (R style indexing)}
}
\value{
true if stable, false if not
}
\description{
Check if a roommate matching is stable
}

