% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{set.column.major}
\alias{set.column.major}
\title{Store preferences in column major order}
\usage{
set.column.major()
}
\description{
After calling this functions, all preferences should be stored in column major
order. This is the default.
}

