% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{set.row.major}
\alias{set.row.major}
\title{Store preference in row major order}
\usage{
set.row.major()
}
\description{
After calling this functions, all preferences should be stored in row major
order.
}

