% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stableRoommateMatching}
\alias{stableRoommateMatching}
\title{Computes a stable roommate matching}
\usage{
stableRoommateMatching(pref)
}
\arguments{
\item{pref}{A matrix with agent's cardinal preferences. Column i is agent i's preferences.}
}
\value{
A list with the matchings made.
}
\description{
This function computes the Irving (1985) algorithm.
}

