% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toptradingcycle.R
\name{toptrading.checkStability}
\alias{toptrading.checkStability}
\title{Check if a one-sided matching for the top trading cycle algorithm is stable}
\usage{
toptrading.checkStability(pref, matchings)
}
\arguments{
\item{pref}{is a matrix with ordinal rankings of the participants}

\item{matchings}{is an nx1 matrix encoding who is matched to whom using
R style indexing}
}
\value{
true if the matching is stable, false otherwise
}
\description{
Check if a one-sided matching for the top trading cycle algorithm is stable
}

