% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{validateInputsOneSided}
\alias{validateInputsOneSided}
\title{Input validation for one-sided markets}
\usage{
validateInputsOneSided(pref = NULL)
}
\arguments{
\item{pref}{is an (n-1)xn matrix, with each column representing the preferences
of a particular agent.}
}
\description{
This function parses and validates the arguments for one sided preferences
for the function onesided.
}

