% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkStabilityRoommate}
\alias{checkStabilityRoommate}
\title{Check if a two-sided matching is stable}
\usage{
checkStabilityRoommate(pref, matchings)
}
\arguments{
\item{pref}{is a matrix with ordinal rankings of the participants}

\item{matchings}{is an nx1 matrix encoding who is matched to whom}
}
\value{
true if the matching is stable, false otherwise
}
\description{
This function checks if a given matching is stable for a particular set of
preferences. This function can check if a given check one-to-one,
one-to-many, or many-to-one matching is stable.
}

