% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_wrapper_irving}
\alias{cpp_wrapper_irving}
\title{Computes a stable roommate matching}
\usage{
cpp_wrapper_irving(pref)
}
\arguments{
\item{pref}{A matrix with agent's cardinal preferences. Column i is agent i's preferences. C++ indexing}
}
\value{
A vector with the matchings made. Unmatched agents are 'matched' to N. C++ indexing
}
\description{
This function computes the Irving (1985) algorithm for finding
a stable matching in a one-sided matching market. Note that neither
existence nor uniqueness is guaranteed, this algorithm finds one
matching, not all of them. If no matching exists, returns 0.
}

