% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_wrapper_ttc}
\alias{cpp_wrapper_ttc}
\title{Computes the top trading cycle algorithm}
\usage{
cpp_wrapper_ttc(pref)
}
\arguments{
\item{pref}{A matrix with agent's cardinal preferences. Column i is agent i's preferences.}
}
\value{
A vector with the matchings made. The matchings are encoded as follows: The first value
in the list is the individual to whom agent 1 will be giving his good, the second value in the list
is the individual to whom agent 2 will be giving his good, etc.
}
\description{
This function uses the top trading cycle algorithm to find a stable trade between agents,
each with some indivisible good, and with preferences over the goods of other agents. Each
agent is matched to one other agent, and matchings are not necessarily two-way. Agents may
be matched with themselves.
}

