% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/galeshapley.R
\name{galeShapley.checkPreferences}
\alias{galeShapley.checkPreferences}
\title{Check if preference order is complete}
\usage{
galeShapley.checkPreferences(pref)
}
\arguments{
\item{pref}{is a matrix with a preference ordering}
}
\value{
a matrix with preference orderings with proper C++ indices or NULL
if the preference order is not complete.
}
\description{
This function checks if a given preference ordering is complete. If needed
it transforms the indices from R indices (starting at 1) to C++ indices
(starting at zero).
}

