% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/galeshapley.R
\name{galeShapley.collegeAdmissions}
\alias{galeShapley.collegeAdmissions}
\title{Uses the Gale-Shapley Algorithm to find solution to the college admission
problem}
\usage{
galeShapley.collegeAdmissions(studentUtils = NULL, collegeUtils = NULL,
  studentPref = NULL, collegePref = NULL, slots = 1,
  studentOptimal = TRUE)
}
\arguments{
\item{studentUtils}{is a matrix with cardinal utilities of the proposing side
of the market. If there are \code{n} students and \code{m} colleges
in the market, then this matrix will be of dimension \code{m} by \code{n}.
The \code{i,j}th element refers to the payoff that student \code{j}
receives from being matched to college \code{i}.}

\item{collegeUtils}{is a matrix with cardinal utilities of the courted side
of the market. If there are \code{n} students and \code{m} colleges
in the market, then this matrix will be of dimension \code{n} by \code{m}.
The \code{i,j}th element refers to the payoff that college \code{j}
receives from being matched to student \code{i}.}

\item{studentPref}{is a matrix with the preference order of the students
(only required when \code{studentUtils} is not provided). If there are
\code{n} students and \code{m} colleges in the market, then this matrix
will be of dimension \code{m} by \code{n}. The \code{i,j}th element refers
to the ID of student \code{j}'s \code{i}th most favorite college.}

\item{collegePref}{is a matrix with the preference order of the colleges
(only required when \code{collegeUtils} is not provided).  If there are
\code{n} students and \code{m} colleges in the market, then this matrix
will be of dimension \code{n} by \code{m}. The \code{i,j}th element refers
to the ID of college \code{j}'s \code{i}th most favorite student}

\item{slots}{is the number of slots per college (this is an integer, i.e. all
colleges have the same number of slots)}

\item{studentOptimal}{is a boolean indicating the proposing side in this
market. If true, students propose and the resulting allocation will be
student-optimal. If false, colleges propose and the resulting allocation
will be college-optimal.}
}
\value{
A list with the successful proposals and engagements:
  \code{proposals} is a vector whose nth element contains the id of the
  reviewer that proposer n is matched to. \code{engagements} is a vector
  whose nth element contains the id of the proposer that reviewer n is
  matched to. \code{unmatched.students} is a vector that lists the ids of
  remaining unmatched students \code{unmatched.colleges} is a vector that
  lists the ids of remaining unmatched colleges (if a college has two slots
  left it will be listed twice)
}
\description{
This function uses the Gale-Shapley algorithm to compute the solution to the
college admissions problem. The function needs some description of
individuals preferences as inputs. That can be in the form of cardinal
utilities or preference orders (or both).
}
\examples{
ncolleges = 10
nstudents = 25
collegeUtils = matrix(runif(ncolleges*nstudents), nrow=nstudents, ncol=ncolleges)
studentUtils = matrix(runif(ncolleges*nstudents), nrow=ncolleges, ncol=nstudents)
results.studentoptimal = galeShapley.collegeAdmissions(studentUtils = studentUtils,
                                                       collegeUtils = collegeUtils,
                                                       slots = 2,
                                                       studentOptimal = TRUE)
results.collegeoptimal = galeShapley.collegeAdmissions(studentUtils = studentUtils,
                                                       collegeUtils = collegeUtils,
                                                       slots = 2,
                                                       studentOptimal = FALSE)
}

