% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{one2one}
\alias{one2one}
\title{Compute the one-to-one matching}
\usage{
one2one(proposerUtils = NULL, reviewerUtils = NULL, proposerPref = NULL,
  reviewerPref = NULL)
}
\arguments{
\item{proposerUtils}{is a matrix with cardinal utilities of the proposing
side of the market}

\item{reviewerUtils}{is a matrix with cardinal utilities of the courted side
of the market}

\item{proposerPref}{is a matrix with the preference order of the proposing
side of the market (only required when \code{proposerUtils} is not provided)}

\item{reviewerPref}{is a matrix with the preference order of the courted side
of the market (only required when \code{reviewerUtils} is not provided)}
}
\value{
A list with the successful proposals and engagements:
\code{proposals} is a vector whose nth element contains the id of the female
that male n is matched to.
\code{engagements} is a vector whose nth element contains the id of the male
that female n is matched to.
\code{single.proposers} is a vector that lists the ids of remaining single
proposers
\code{single.reviewers} is a vector that lists the ids of remaining single
reviewers
}
\description{
This function returns the one-to-one matching. The function needs some
description of individuals preferences as inputs. That can be in the form of
cardinal utilities or preference orders (or both). It is computational most
efficient to provide preference orders for the proposers \code{proposerPref}
and cardinal utilities for the reviewers \code{reviewerPref}.
}

