% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{onesided}
\alias{onesided}
\title{Compute matching for one-sided markets}
\usage{
onesided(pref = NULL, prefUtil = NULL)
}
\arguments{
\item{pref}{An nxn-1 matrix, with each column representing the cardinal
utilities of each agent over matches with the other agents, so that, e.g.,
if element (4, 6) of this matrix is 2, then agent 4 ranks agent 2 6th.}

\item{prefUtil}{An nxn-1 matrix, each column representing ordinal preferences
of each agent over agents 1, 2, ..., i-1, i+1, i+2, ... n.}
}
\value{
A list of length n corresponding to the matchings being made, so that
e.g. if the 4th element is 6 then agent 4 was matched with agent 6.
}
\description{
This function returns a stable roommate matching for a one-sided market
using Irving (1985)'s algorithm. Stable matchings are neither guaranteed
to exist, nor to be unique.
}
\examples{
p = replicate(100, rnorm(99))
results = onesided(prefUtil = p)
}

