% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rankIndex}
\alias{rankIndex}
\title{Rank elements within row of a matrix}
\usage{
rankIndex(sortedIdx)
}
\arguments{
\item{u}{is the input matrix}
}
\value{
a rank matrix
}
\description{
This function returns the rank of each element within each row of a matrix.
The highest element receives the highest rank.
}

