% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roommate.R
\name{roommate.checkPreferences}
\alias{roommate.checkPreferences}
\title{Check if preference order for a one-sided market is complete.}
\usage{
roommate.checkPreferences(pref)
}
\arguments{
\item{pref}{is a matrix with a preference ordering for a one-sided market.
If necessary transforms the indices from R indices (starting at 1) to C++
indices (starting at 0).}
}
\value{
a matrix with preference orderings with proper C++ indices or NULL
if the preference order is not complete.
}
\description{
Check if preference order for a one-sided market is complete.
}

