% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roommate.R
\name{roommate.validate}
\alias{roommate.validate}
\title{Input validation for one-sided markets}
\usage{
roommate.validate(pref = NULL, utils = NULL)
}
\arguments{
\item{pref}{is an n-1xn matrix, with each row representing an ordinal ranking.}

\item{utils}{if an n-1xn matrix, with each row representing the cardinal preferences
of the agents.}
}
\value{
The validated inputs. C++ style indexing.
}
\description{
This function parses and validates the arguments for one sided preferences
for the function onesided. If it uses R-style indexing (i.e., beginning at
1), then it re-numbers the preference matrix to use C++ style indexing.
}

