% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sortIndex}
\alias{sortIndex}
\title{Sort indices of a matrix within a column}
\usage{
sortIndex(u)
}
\arguments{
\item{u}{is the input matrix}
}
\value{
a matrix with sorted indicies
}
\description{
Within each column of a matrix, this function returns the indices of each
element in descending order
}

