% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stableRoommateMatching}
\alias{stableRoommateMatching}
\title{Computes a stable roommate matching}
\usage{
stableRoommateMatching(pref)
}
\arguments{
\item{pref}{A matrix with agent's cardinal preferences. Column i is agent i's preferences.}
}
\value{
A list with the matchings made. Unmatched agents are 'matched' to N.
}
\description{
This function computes the Irving (1985) algorithm for finding
a stable matching in a one-sided matching market. Note that neither
existence nor uniqueness is guaranteed, this algorithm finds one
matching, not all of them. If no matching exists, returns 0.
}

