% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roommate.R
\name{validateInputsOneSided}
\alias{validateInputsOneSided}
\title{Input validation for one-sided markets}
\usage{
validateInputsOneSided(pref = NULL, utils = NULL)
}
\arguments{
\item{pref}{is an n-1xn matrix, with each row representing an ordinal ranking.}

\item{utils}{if an n-1xn matrix, with each row representing the cardinal preferences
of the agents.}
}
\value{
The validated inputs, ready to be sent to C++ code.
}
\description{
This function parses and validates the arguments for one sided preferences
for the function onesided. If it uses R-style indexing (i.e., beginning at
1), then it re-numbers the preference matrix to use C++ style indexing.
}

