% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchingR.R
\name{validateInputsOneSided}
\alias{validateInputsOneSided}
\title{Input validation for one-sided markets}
\usage{
validateInputsOneSided(pref = NULL, prefUtil = NULL)
}
\arguments{
\item{pref}{is an nx(n-1) matrix, with each row representing an ordinal ranking.}

\item{prefUtil}{Cardinal preferences of the agents.
of a particular agent.}
}
\description{
This function parses and validates the arguments for one sided preferences
for the function onesided.
}

